backend-django/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes -o requirements.txt
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.2
    # via requests
dj-database-url==3.0.1
    # via backend-django
django==5.2.4
    # via
    #   backend-django
    #   dj-database-url
    #   django-cors-headers
    #   django-filter
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==4.7.0
    # via backend-django
django-filter==25.1
    # via backend-django
django-shortcuts==1.6
    # via backend-django
django-stubs==5.2.2
    # via djangorestframework-stubs
django-stubs-ext==5.2.2
    # via django-stubs
djangorestframework==3.16.0
    # via
    #   backend-django
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-jwt==1.11.0
    # via backend-django
djangorestframework-simplejwt==5.5.1
    # via backend-django
djangorestframework-stubs==3.16.1
drf-yasg==1.21.10
    # via backend-django
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-yasg
mysqlclient==2.2.7
    # via backend-django
packaging==25.0
    # via drf-yasg
pyjwt==1.7.1
    # via
    #   djangorestframework-jwt
    #   djangorestframework-simplejwt
python-dotenv==1.1.1
    # via backend-django
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.2
    # via drf-yasg
requests==2.32.4
    # via
    #   backend-django
    #   djangorestframework-stubs
sqlparse==0.5.3
    # via django
types-pyyaml==6.0.12.20250516
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.4.20250611
    # via djangorestframework-stubs
typing-extensions==4.14.1
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
tzdata==2025.2 ; sys_platform == 'win32'
    # via django
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
gunicorn==21.2.0
